Update job submission for Tl

+++ NEXT100/job/NEXT100_Kr83m.sh
@@ -55,11 +55,15 @@ city hypathia hypathiaTemplate.conf -i NEXT100_Kr83m_detsim_${JOBID}.h5   -o NEX
 city dorothea dorotheaTemplate.conf -i NEXT100_Kr83m_hypathia_${JOBID}.h5 -o NEXT100_Kr83m_dorothea_${JOBID}.h5
 
 rm NEXT100_Kr83m.h5
-rm NEXT100_Kr83m_detsim_${JOBID}.h5
-rm NEXT100_Kr83m_hypathia_${JOBID}.h5
 rm *LT*
 rm *PSF*
 rm *map*
+
+# Only keep first 1000 files for validation purposes
+if (( JOBID > 1000 )); then
+    rm NEXT100_Kr83m_detsim_${JOBID}.h5
+    rm NEXT100_Kr83m_hypathia_${JOBID}.h5
+fi
 
 ls -ltrh
